move contact-controller to controllers folder
